+++ CHANGELOG.rst
@@ -4,8 +4,13 @@ CHANGELOG
 UNRELEASED
 ----------
 
-* 🐛 fix: create_scene_instance now returns a Node instead of a PackedScene
-* 🎉 feat: your scenes now save in `addons/scene_manager/scenes.gd` and you do not need to have a file near every export
+
+2.0.0 (2022-03-29)
+------------------
+
+* fix: create_scene_instance now returns a Node instead of a PackedScene
+* feat: your scenes now save in `addons/scene_manager/scenes.gd` and you do not need to have a file near every export of your games
+* note: this version does not support .json files of previous versions
 
 1.4.0 (2022-03-25)
 ------------------